test(run-cloud): register lifecycle coverage
